pyproject.toml
[build-system]
requires = ["setuptools>=72", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "payerai-gpt"
version = "0.1.0"
description = "Healthcare payer AI platform for prior authorization, claims intelligence, and coverage determination."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "Proprietary" }

dependencies = [
    "azure-ai-projects>=1.0.0",
    "azure-identity>=1.19.0",
    "openai>=1.54.0",
    "chromadb>=0.5.0",
    "pypdf>=4.3.0",
    "python-docx>=1.1.0",
    "pydantic>=2.9.0",
    "anthropic>=0.40.0",
    "claude-code-sdk>=0.0.1",
    "mcp[cli]>=1.0.0",
    "tiktoken>=0.8.0",
    "numpy>=1.26.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "respx>=0.21.0",
    "httpx>=0.27.0",
]
dev = [
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

# ── pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

# ── ruff ──────────────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
